netram_starter_kit/services/ai-service/main.py
from fastapi import FastAPI, UploadFile, File, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import dotenv
import os
# import openai # Uncomment if using OpenAI
# import google.generativeai as genai # Uncomment if using Gemini

dotenv.load_dotenv()

app = FastAPI(
    title="Netram AI Service",
    description="Provides AI/ML capabilities for the Netram Eye Health Platform.",
    version="0.1.0"
)

# CORS Configuration (adjust as needed for your frontend origins)
origins = [
    "http://localhost", # Allow local development for web-client
    "http://localhost:3000", # Default Next.js dev port
    "http://localhost:5173", # Default Vite dev port
    # Add your deployed frontend URLs here
]

app.add_middleware(
    CORSMiddleware,
    allow_origins=origins,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# --- Configure AI Models ---
# OpenAI (GPT) Example
# openai_api_key = os.getenv("OPENAI_API_KEY")
# if openai_api_key:
#     openai.api_key = openai_api_key
#     print("OpenAI API key loaded.")
# else:
#     print("Warning: OPENAI_API_KEY not found in .env file.")

# Google Gemini Example
# google_gemini_api_key = os.getenv("GOOGLE_GEMINI_API_KEY")
# if google_gemini_api_key:
#     genai.configure(api_key=google_gemini_api_key)
#     print("Google Gemini API key loaded.")
# else:
#     print("Warning: GOOGLE_GEMINI_API_KEY not found in .env file.")


class DiagnosisRequest(BaseModel):
    symptoms: str
    patient_history: str = None

class DiagnosisResponse(BaseModel):
    possible_conditions: list[str]
    confidence_scores: list[float]
    recommendations: str

class TranscriptionRequest(BaseModel):
    audio_data_url: str # Assuming audio is sent as a data URL

class TranscriptionResponse(BaseModel):
    transcript: str

class ImageAnalysisResponse(BaseModel):
    findings: list[str]
    summary: str

@app.get("/")
async def read_root():
    return {"message": "Netram AI Service is running!"}

@app.post("/ai-diagnosis", response_model=DiagnosisResponse)
async def diagnose_symptoms(request: DiagnosisRequest):
    """
    Suggest possible conditions based on symptoms and patient history.
    (Placeholder - implement actual AI model call here)
    """
    print(f"Received diagnosis request: {request.symptoms}")
    # Example: Call Gemini or GPT-4 with a prompt
    # prompt = f"A patient presents with symptoms: {request.symptoms}. Patient history: {request.patient_history}. What are the possible conditions and recommendations?"
    # response_text = call_your_llm(prompt) # Replace with actual LLM call
    
    # Placeholder response
    return DiagnosisResponse(
        possible_conditions=["Common Cold", "Allergic Rhinitis"],
        confidence_scores=[0.7, 0.6],
        recommendations="Consult a doctor for accurate diagnosis. Consider rest and hydration."
    )

@app.post("/voice-transcriber", response_model=TranscriptionResponse)
async def transcribe_audio(request: TranscriptionRequest):
    """
    Transcribe doctor's speech to text for EMR notes.
    (Placeholder - implement actual speech-to-text model call here)
    """
    print(f"Received transcription request for audio data.")
    # Example: Decode data URL, send to Whisper API or Gemini speech-to-text
    # transcript = call_speech_to_text_api(request.audio_data_url)

    # Placeholder response
    return TranscriptionResponse(transcript="Patient complains of slight blurriness in the left eye for the past two days. No pain reported.")

@app.post("/image-analyzer", response_model=ImageAnalysisResponse)
async def analyze_image(image_file: UploadFile = File(...)):
    """
    Analyze fundus/OCT images for anomalies.
    (Placeholder - implement actual image analysis model call here)
    """
    contents = await image_file.read()
    print(f"Received image for analysis: {image_file.filename}, size: {len(contents)} bytes")
    # Example: Send image_file.file (BytesIO object) or contents to a vision model (Gemini Vision, TensorFlow model)
    # analysis_result = call_image_analysis_model(contents)

    # Placeholder response
    return ImageAnalysisResponse(
        findings=["Drusen detected in macula", "Optic disc appears normal"],
        summary="Signs consistent with early Age-related Macular Degeneration (AMD). Recommend follow-up OCT."
    )

# Add more endpoints for 'chat-emr', 'summary-writer', etc.

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=int(os.getenv("PORT_AI_SERVICE", 8000)))
```